btxld: build errors with clang
/code/illumos-gate/usr/src/cmd/boot/common/boot_utils.c:78:11: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                if (ret < size) {
                    ~~~ ^ ~~~~
/code/illumos-gate/usr/src/cmd/boot/common/boot_utils.c:107:11: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                if (ret < size) {
                    ~~~ ^ ~~~~
2 errors generated.

/code/illumos-gate/usr/src/cmd/boot/common/mboot_extra.c:52:16: error: comparison of integers of different signs: 'int' and 'uint32_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        for (i = 0; i < size; i += sizeof (uint32_t))
                    ~ ^ ~~~~
/code/illumos-gate/usr/src/cmd/boot/common/mboot_extra.c:77:16: error: comparison of integers of different signs: 'int' and 'uint32_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        for (i = 0; i < boundary; i += 4, iter++) {
                    ~ ^ ~~~~~~~~
2 errors generated.

btxld.c:349:20: error: comparison of integers of different signs: 'off_t' (aka 'long') and 'unsigned int' [-Werror,-Wsign-compare]
    if (sb.st_size > UINT32_MAX)
        ~~~~~~~~~~ ^ ~~~~~~~~~~
1 error generated.

+++ usr/src/cmd/boot/common/boot_utils.c
@@ -56,7 +56,7 @@ boot_gdebug(const char *funcname, char *format, ...)
 int
 write_out(int fd, void *buffer, size_t size, off_t off)
 {
-	int		ret;
+	ssize_t		ret;
 	char		*buf = buffer;
 
 	if (size % SECTOR_SIZE != 0)
@@ -75,7 +75,7 @@ write_out(int fd, void *buffer, size_t size, off_t off)
 			else
 				return (BC_ERROR);
 			}
-		if (ret < size) {
+		if ((size_t)ret < size) {
 			size -= ret;
 			off += ret;
 			buf += ret;
@@ -89,7 +89,7 @@ write_out(int fd, void *buffer, size_t size, off_t off)
 int
 read_in(int fd, void *buffer, size_t size, off_t off)
 {
-	int		ret;
+	ssize_t		ret;
 	char		*buf = buffer;
 
 	if (size % SECTOR_SIZE != 0)
@@ -104,7 +104,7 @@ read_in(int fd, void *buffer, size_t size, off_t off)
 			else
 				return (BC_ERROR);
 			}
-		if (ret < size) {
+		if ((size_t)ret < size) {
 			size -= ret;
 			off += ret;
 			buf += ret;
